Fixed two defects and modified how items are identified
The defects were due to multiple edits and changing of variable names due to those edits.  I also saw a comment highlighting the fact that item_name is not as unique as item_id.  Based on the example it appears that item_id was the desired value (183 item_ids items versus 179 item_names.  However, depending on the business case, item_name may be preferred.  It could have been a simple data load issue that caused numbering issues and the source data has not yet been cleaned to address the data load error.  I would raise this concern to the person that supplied the data and the person providing the business requirements justifying the analysis.