Cleaned up file
Based on previous HW comments I realized the goal of the file was to show the final code versus showing the work it took to get to the final code.  So I removed most of my comments and focused the new comments on explaining the code versus explaining how I arrived at the solution.